uses math/bits for trailing zeros
